\name{Cedar2020L1}
\alias{Cedar2020L1}
\title{Cedar2020L1 Data }
\usage{
This is the most recent flood event that started in January 2020. Used to add another event to the concurrent model to predict the 2015 event. The specific times for this event were choses using the L1 difference method and then it was downloaded from USGS website.
}
\description{
	Acquired from:
https://waterdata.usgs.gov/sc/nwis/uv?site_no=02169672
It goes from 1/29/20 22:00 3/22/20,13:30
}
\examples{
N/A
}