\name{EntireCongareeHt}
\alias{EntireCongareeHt}
\title{EntireCongareeHt Data}
\usage{
This is raw Congaree River Height data from 1/1/95 to 9/30/19. There are 799,850 records in 15 minute intervals. 
}
\description{
The data is from:
Jason M. Fine | U.S. Geological Survey
Acting Associate Director for Data - Caribbean Florida Water Science Center 
Assistant Director for Data - North Carolina - South Atlantic Water Science Center 
On February 4, 2020.
}
\examples{
N/A
}
