\name{Error.fRegress}
\alias{Error.fRegress}
\title{Error.fRegress Function}
\usage{
Calculates the Mean Squared Error across the curve using the given predictor and response datasets to predict the response curve and then compares the prediction to the true response curve that is known.
}
\description{
A simple MSE across each index. This function is typically not used by itself, rather, it is used inside the bestEst function to find the optimal number of basis functions to use when using the FRegress function.}
\examples{
NA #To be used with the bestEst function.
}