\name{FinalXtL1star}
\alias{FinalXtL1star}
\title{FinalXtL1star Data}
\usage{
This data should be used in correspondence with FinalYtL1star in the predict.FRgressNormTest function. 
}
\description{
This is the Congaree River heights for 10 distinct flood events between 1995 and 2020 using the L1* selection method.
}
\examples{
N/A
}