\name{FinalYtL1star}
\alias{FinalYtL1star}
\title{FinalYtL1star Data}
\usage{
This data should be used in correspondence with FinalXtL1star in the predict.FRgressNormTest function.
}
\description{
This is the corresponding Cedar Creek heights for 10 distinct flood events between 1995 and 2020 using the L1* selection method.
}
\examples{
N/A
}